# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API для управления автомобилями",
        "description": "Документация API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/car-models": {
            "get": {
                "tags": [
                    "Car Models"
                ],
                "summary": "Список моделей автомобилей",
                "operationId": "e3b490883632ae064d1de5ddac65cc7e",
                "responses": {
                    "200": {
                        "description": "Список моделей",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CarModelModel"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Car Models"
                ],
                "summary": "Создать модель автомобиля",
                "operationId": "3484efe17cc38d2cbf85e61aaaee781a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "brand_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Camry"
                                    },
                                    "brand_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Модель создана",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/CarModelModel"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/car-models/{id}": {
            "get": {
                "tags": [
                    "Car Models"
                ],
                "summary": "Показать модель автомобиля",
                "operationId": "64a838d8b35a648d96ece48c74a9d740",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID модели",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные модели",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/CarModelModel"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Car Models"
                ],
                "summary": "Удалить модель автомобиля",
                "operationId": "891983dc0d18b7d798f7ca2459643ab8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID модели",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Модель удалена"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Car Models"
                ],
                "summary": "Обновить модель автомобиля",
                "operationId": "1f44fdf2b27dec69bb9913a394875b65",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID модели",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "brand_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Camry"
                                    },
                                    "brand_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Модель обновлена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/CarModelModel"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "CarModelModel": {
                "title": "Car Model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "brand_id": {
                        "type": "integer",
                        "example": 10
                    },
                    "name": {
                        "type": "string",
                        "example": "Vesta"
                    }
                },
                "type": "object"
            }
        },
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}